// orion-knowledge/web/admin/src/request/types.ts
export enum DomainStatPageScene {
  StatPageSceneWelcome = 1,
  StatPageSceneNodeDetail = 2,
  StatPageSceneChat = 3,
  StatPageSceneLogin = 4,
}
// ...
export enum DomainAppType {
  AppTypeWeb = 1,
  AppTypeWidget = 2,
  AppTypeDingTalkBot = 3,
  AppTypeFeishuBot = 4,
  AppTypeWechatBot = 5,
  AppTypeWechatServiceBot = 6,
  AppTypeDisCordBot = 7,
  AppTypeWechatOfficialAccount = 8,
  AppTypeOpenAIAPI = 9,
  AppTypeWecomAIBot = 10,
  AppTypeLarkBot = 11,
  AppTypeMcpServer = 12,
}
// ...
export interface DomainAuthUserInfo {
  avatar_url?: string;
  email?: string;
  username?: string;
}
// ...
export interface DomainBrowserCount {
  count?: number;
  name?: string;
}
// ...
export interface DomainHotBrowser {
  browser?: DomainBrowserCount[];
  os?: DomainBrowserCount[];
}

export interface DomainHotPage {
  count?: number;
  node_id?: string;
  node_name?: string;
  scene?: DomainStatPageScene;
}

export interface DomainHotRefererHost {
  count?: number;
  referer_host?: string;
}

export interface DomainIPAddress {
  city?: string;
  country?: string;
  ip?: string;
  province?: string;
}

export interface DomainImgTextConfig {
  item?: {
    desc?: string;
    name?: string;
    url?: string;
  };
  title?: string;
  type?: string;
}

export interface DomainInstantCountResp {
  count?: number;
  time?: string;
}

export interface DomainInstantPageResp {
  created_at?: string;
  info?: DomainAuthUserInfo;
  ip?: string;
  ip_address?: DomainIPAddress;
  node_id?: string;
  node_name?: string;
  scene?: DomainStatPageScene;
  user_id?: number;
}
// ...
export interface DomainPWResponse {
  code?: number;
  data?: unknown;
  message?: string;
  success?: boolean;
}
// ...
export interface DomainResponse {
  data?: unknown;
  message?: string;
  success?: boolean;
}
// ...
export interface V1StatConversationDistributionResp {
  app_type?: DomainAppType;
  count?: number;
}

export interface V1StatCountResp {
  conversation_count?: number;
  ip_count?: number;
  page_visit_count?: number;
  session_count?: number;
}
// ...
export interface GetApiV1StatBrowsersParams {
  day?: 1 | 7 | 30 | 90;
  kb_id: string;
}

export interface GetApiV1StatConversationDistributionParams {
  day?: 1 | 7 | 30 | 90;
  kb_id: string;
}

export interface GetApiV1StatCountParams {
  day?: 1 | 7 | 30 | 90;
  kb_id: string;
}

export interface GetApiV1StatGeoCountParams {
  day?: 1 | 7 | 30 | 90;
  kb_id: string;
}

export interface GetApiV1StatHotPagesParams {
  day?: 1 | 7 | 30 | 90;
  kb_id: string;
}

export interface GetApiV1StatInstantCountParams {
  kb_id: string;
}

export interface GetApiV1StatInstantPagesParams {
  kb_id: string;
}

export interface GetApiV1StatRefererHostsParams {
  day?: 1 | 7 | 30 | 90;
  kb_id: string;
}